Use lambda_stmt for recorder queries and migrate them to queries module (#71219)
